add playlist feature

+++ src/services/postgres/PlaylistsService.js
@@ -1,23 +1,25 @@
 const { Pool } = require('pg');
 const { nanoid } = require('nanoid');
 const { mapPlaylistDBToModel } = require('../../utils/playlistModel');
+const { mapMusicDBToModel } = require('../../utils/musicModel');
 const NotFoundError = require('../../exceptions/NotFoundError');
 const InvariantError = require('../../exceptions/InvariantError');
+const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistsService {
   constructor() {
     this._pool = new Pool();
   }
 
-  async verifyPlaylistAccess(noteId, userId) {
+  async verifyPlaylistAccess(playlistId, userId) {
     try {
-      await this.verifyPlaylistOwner(noteId, userId);
+      await this.verifyPlaylistOwner(playlistId, userId);
     } catch (error) {
       if (error instanceof NotFoundError) {
         throw error;
       }
       try {
-        await this._collaborationService.verifyCollaborator(noteId, userId);
+        await this._collaborationService.verifyCollaborator(playlistId, userId);
       } catch {
         throw error;
       }
@@ -26,33 +28,40 @@ class PlaylistsService {
 
   async verifyPlaylistOwner(id, owner) {
     const query = {
-      text: 'SELECT * FROM playlist WHERE id = $1',
+      text: 'SELECT * FROM playlists WHERE id = $1',
       values: [id],
     };
     const result = await this._pool.query(query);
     if (!result.rows.length) {
       throw new NotFoundError('Playlist tidak ditemukan');
     }
-    const note = result.rows[0];
-    if (note.owner !== owner) {
+    const playlist = result.rows[0];
+    if (playlist.owner !== owner) {
       throw new AuthorizationError('Anda tidak berhak mengakses resource ini');
     }
   }
 
-  async addSongToPlaylists(id,songId) {
+  async addSongToPlaylists(playlistId,songId,owner) {
+    await this.verifyPlaylistAccess(playlistId,owner)
 
-    const query = {
-      text: 'INSERT INTO playlists_songs VALUES($1,$2) RETURNING id',
-      values: [id,songId],
+    const resultSong = await this._pool.query("SELECT * FROM song WHERE id = $1",[songId]);
+    if (!resultSong.rows.length) {
+      throw new NotFoundError('Lagu tidak ditemukan');
+    }else{
+      const id = nanoid(16);
+      const query = {
+      text: 'INSERT INTO playlist_songs VALUES($1, $2, $3) RETURNING id',
+      values: [id,playlistId,songId],
     };
 
     const result = await this._pool.query(query);
 
     if (!result.rows[0].id) {
-      throw new InvariantError('Playlist gagal ditambahkan');
+      throw new InvariantError('Lagu gagal ditambahkan ke playlist');
     }
 
     return result.rows[0].id;
+    }
   }
 
   async addPlaylists({ name, owner }) {
@@ -74,7 +83,18 @@ class PlaylistsService {
 
   async getPlaylists(owner){
 
-    const result = await this._pool.query("SELECT id, name, owner as username FROM playlists WHERE owner = $1",[owner]);
+    const result = await this._pool.query(`SELECT
+    playlists."id", 
+    playlists."name", 
+    users.username
+  FROM
+    playlists
+    INNER JOIN
+    users
+    ON 
+      playlists."owner" = users."id"
+  WHERE
+    playlists."owner" =  $1`,[owner]);
 
     if (!result.rows.length) {
       throw new NotFoundError('Playlists tidak ditemukan');
@@ -82,25 +102,45 @@ class PlaylistsService {
     return result.rows.map(mapPlaylistDBToModel)
   }
 
-  async getPlaylistsById(id) {
-    const query = {
-      text: 'SELECT * FROM playlists WHERE id = $1',
-      values: [id],
-    };
-    const result = await this._pool.query(query);
+  async getPlaylistsById(id,owner) {
+    await this.verifyPlaylistAccess(id,owner)
+
     const queryMusic = {
-      text: 'SELECT id,title,performer FROM song WHERE "playlistsId" = $1',
+      text: `SELECT
+      playlists."id" as playlistid, 
+      playlists."name", 
+      song."id", 
+      song.title, 
+      song.performer, 
+      users.username
+    FROM
+      playlists
+      INNER JOIN
+      playlist_songs
+      ON 
+        playlists."id" = playlist_songs.playlist_id
+      INNER JOIN
+      song
+      ON 
+        playlist_songs.song_id = song."id"
+      INNER JOIN
+      users
+      ON 
+        playlists."owner" = users."id"
+    WHERE
+      playlists."id" =  $1`,
       values: [id],
     };
     const resultMusic = await this._pool.query(queryMusic);
-    if (!result.rows.length) {
-      throw new NotFoundError('Playlists tidak ditemukan');
+    if (!resultMusic.rows.length) {
+      throw new NotFoundError('Lagu tidak ditemukan');
     }
+
     const res = {
-      "Playlists": {
-        "id": result.rows[0].id,
-        "name": result.rows[0].name,
-        "year": result.rows[0].year,
+      "playlist": {
+        "id": resultMusic.rows[0].playlistid,
+        "name": resultMusic.rows[0].name,
+        "username": resultMusic.rows[0].username,
         "songs": resultMusic.rows.map(mapMusicDBToModel)
       }
     };
@@ -122,7 +162,25 @@ class PlaylistsService {
     }
   }
 
-  async deletePlaylistsById(id) {
+  async deleteSongPlaylistsById(id, songId, owner) {
+    await this.verifyPlaylistAccess(id,owner)
+
+    const query = {
+      text: 'DELETE FROM playlist_songs WHERE song_id = $1 RETURNING id',
+      values: [songId],
+    };
+
+    const result = await this._pool.query(query);
+
+    if (!result.rows.length) {
+      throw new NotFoundError('Lagu gagal dihapus dari playlist. Id tidak ditemukan');
+    }
+  }
+  async deletePlaylistsById(id, owner) {
+    await this.verifyPlaylistAccess(id,owner)
+
+    await this._pool.query('DELETE FROM playlist_songs WHERE playlist_id = $1 RETURNING id',[id]);
+
     const query = {
       text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
       values: [id],
@@ -131,7 +189,7 @@ class PlaylistsService {
     const result = await this._pool.query(query);
 
     if (!result.rows.length) {
-      throw new NotFoundError('Playlists gagal dihapus. Id tidak ditemukan');
+      throw new NotFoundError('Playlist gagal dihapus. Id tidak ditemukan');
     }
   }
 }
